# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CCS API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cat/case-details": {
            "post": {
                "tags": [
                    "CAT"
                ],
                "summary": "Get CAT case details by case number",
                "operationId": "51a9b69c90406fa2436b67d9563b1e44",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "catschemaId",
                                    "casetypeId",
                                    "caseNo",
                                    "caseYear"
                                ],
                                "properties": {
                                    "catschemaId": {
                                        "description": "Location code (e.g., 117 for Jammu, 119 for Srinagar)",
                                        "type": "string",
                                        "example": "117"
                                    },
                                    "casetypeId": {
                                        "description": "Case type ID (e.g., 1 for Original Application)",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "caseNo": {
                                        "description": "Case number",
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "caseYear": {
                                        "description": "Case year",
                                        "type": "string",
                                        "example": "2024"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "data": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "error"
                                                },
                                                "error": {
                                                    "type": "string",
                                                    "example": "Invalid API response or no data found"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/cat/daily-orders": {
            "post": {
                "tags": [
                    "CAT"
                ],
                "summary": "Get CAT daily orders by case number",
                "operationId": "21a63b4c4c0cc038060c925ee2dc1df2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "catschemaId",
                                    "casetypeId",
                                    "caseNo",
                                    "caseYear"
                                ],
                                "properties": {
                                    "catschemaId": {
                                        "description": "Location code (e.g., 117 for Jammu, 119 for Srinagar)",
                                        "type": "string",
                                        "example": "117"
                                    },
                                    "casetypeId": {
                                        "description": "Case type ID (e.g., 1 for Original Application)",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "caseNo": {
                                        "description": "Case number",
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "caseYear": {
                                        "description": "Case year",
                                        "type": "string",
                                        "example": "2024"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "error"
                                                },
                                                "error": {
                                                    "type": "string",
                                                    "example": "Invalid API response or no data found"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
...
}